Feat: 관리자 유저, 제품 관리 기능에서 데이터 수정 혹은 삭제시 사용자의 확인을 받는 로직을 추가.
Refactor: 변경된 방향성대로 회원 기능 코드 개선 작업 실행.
(공통 함수 부분을 삭제하고, 서비스 레이어 분리 작업만 실행.)
(되돌린 이후 기능들의 정상 동작을 확인함.)

+++ frontend/src/components/admin/InventoryManagement.tsx
@@ -178,7 +178,9 @@ const InventoryManagement: React.FC = () => {
                                 <TableCell>
                                     <ActionButton
                                         className="edit"
-                                        onClick={() => openEditModal(product)}
+                                        onClick={() => {
+                                            openEditModal(product);
+                                        }}
                                     >
                                         수정
                                     </ActionButton>
@@ -188,9 +190,15 @@ const InventoryManagement: React.FC = () => {
                                             if (
                                                 product.product_id !== undefined
                                             ) {
-                                                handleDeleteProduct(
-                                                    product.product_id
-                                                );
+                                                if (
+                                                    window.confirm(
+                                                        '상품을 삭제하시겠습니까?'
+                                                    )
+                                                ) {
+                                                    handleDeleteProduct(
+                                                        product.product_id
+                                                    );
+                                                }
                                             }
                                         }}
                                     >
